agents/enterprise/multiple-customer-support/production-chatbot/backend/requirements.txt
# AWS SDK
boto3==1.35.0
botocore==1.35.0

# Data validation
pydantic==2.10.6
pydantic-settings==2.7.1

# Configuration
pyyaml==6.0.2

# Utilities
python-dateutil==2.9.0
